backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
djangorestframework-simplejwt==5.2.2
python-dotenv==1.0.0
dropbox==12.0.2
django-import-export==4.1.1
tablib==3.5.0
openpyxl==3.1.5
xlrd==2.0.1
gunicorn==23.0.0
whitenoise==6.9.0
psycopg2-binary==2.9.9
requests==2.32.5
